# docs/swagger.yaml
basePath: /
definitions:
  badresponse.ResponseError:
    properties:
      reason:
        type: string
      time:
        type: string
    type: object
...
  dto.UserInfo:
    properties:
      email:
        example: email@qq.com
        type: string
      id:
        example: xxxxxxxxxxxxx==
        type: string
      name:
        example: account name
        type: string
      roles:
        example:
        - ordinaryUser
        items:
          type: string
        type: array
      settings:
        additionalProperties:
          type: string
        example:
          '''avatar''': ' ''www.avatar.com/account_name'''
          '''site''': '''www.limfx.com''(public data)'
        type: object
    type: object
  dto.UserInfoBrief:
    properties:
      email:
        example: email@qq.com
        type: string
      id:
        example: xxxxxxxxxxxxx==
        type: string
      name:
        example: account name
        type: string
      settings:
        additionalProperties:
          type: string
        example:
          '''avatar''': ' ''www.avatar.com/account_name'''
          '''site''': '''www.limfx.com''(public data)'
        type: object
    type: object
  model.AddAccount:
    properties:
      email:
        example: email@mo2.com
        type: string
      password:
        example: p@ssword
        type: string
      userName:
        example: account name
        type: string
    type: object
  model.AddAccountRole:
    properties:
      id:
        example: xxxxxxxxxxxxx==
        type: string
      roles:
        items:
          type: string
        type: array
      super_key:
        example: special
        type: string
    type: object
...
  model.DeleteAccount:
    properties:
      email:
        example: email@mo2.com
        type: string
      password:
        example: p@ssword
        type: string
    type: object
...
  model.LoginAccount:
    properties:
      password:
        example: p@ssword
        type: string
      userNameOrEmail:
        example: account name/email@mo2.com
        type: string
    type: object
...
host: localhost:5001
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a Mo2 server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Mo2
  version: "1.0"
paths:
  /api/accounts:
    delete:
      consumes:
      - application/json
      description: delete by path
      parameters:
      - description: delete account info
        in: body
        name: info
        required: true
        schema:
          $ref: '#/definitions/model.DeleteAccount'
      produces:
      - application/json
      responses:
        "202":
          description: ""
        "204":
          description: ""
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/badresponse.ResponseError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/badresponse.ResponseError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/badresponse.ResponseError'
      summary: delete Blog
      tags:
      - accounts
    post:
      consumes:
      - application/json
      description: add by json account
      parameters:
      - description: add new account info
        in: body
        name: account
        required: true
        schema:
          $ref: '#/definitions/model.AddAccount'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.UserInfo'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/badresponse.ResponseError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/badresponse.ResponseError'
      summary: Add an account
      tags:
      - accounts
    put:
      consumes:
      - application/json
      description: 通过id获取已有用户，验证身份。并将name的修改与setting的修改应用
      parameters:
      - description: id必须，可修改name/settings
        in: body
        name: account
        required: true
        schema:
          $ref: '#/definitions/dto.UserInfoBrief'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.UserInfo'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/badresponse.ResponseError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/badresponse.ResponseError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/badresponse.ResponseError'
      summary: 修改名称（唯一用于登录）/偏好设置
      tags:
      - accounts
  /api/accounts/detail/{id}:
    get:
      consumes:
      - application/json
      description: get string by ID；若id为空，返回所有用户信息
      operationId: get-string-by-int
      parameters:
      - description: Account ID
        in: path
        name: id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.UserInfo'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/badresponse.ResponseError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/badresponse.ResponseError'
      summary: Show account's info
      tags:
      - accounts
  /api/accounts/listBrief:
    get:
      consumes:
      - application/json
      description: from a list of user ids [usage]:/api/accounts/listBrief?id=60223d4042d6febff9f276f0&id=60236866d2a68483adaccc38
      parameters:
      - description: user IDs list
        in: query
        name: id
        required: true
        type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.UserInfoBrief'
            type: array
      summary: List accounts brief info
      tags:
      - accounts
  /api/accounts/login:
    post:
      consumes:
      - application/json
      description: login by json model.LoginAccount and set cookies
      parameters:
      - description: login account
        in: body
        name: account
        required: true
        schema:
          $ref: '#/definitions/model.LoginAccount'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.UserInfo'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/badresponse.ResponseError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/badresponse.ResponseError'
      summary: login an account
      tags:
      - accounts
...
  /api/accounts/role:
    post:
      consumes:
      - application/json
      description: add by json account
      parameters:
      - description: add new account info
        in: body
        name: account
        required: true
        schema:
          $ref: '#/definitions/model.AddAccountRole'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.UserInfo'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/badresponse.ResponseError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/badresponse.ResponseError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/badresponse.ResponseError'
      summary: Add role for an account
      tags:
      - admin
  /api/accounts/verify:
    get:
      consumes:
      - application/json
      description: add by json account
      parameters:
      - description: email@mo2.com
        in: query
        name: email
        required: true
        type: string
      - description: xxxx==sf
        in: query
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "308":
          description: ""
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/badresponse.ResponseError'
      summary: verify an account's email
      tags:
      - accounts
...
swagger: "2.0"